recommend-legacy/requirements.txt
fastapi
uvicorn[standard]
pydantic

# AI - CLIP (추천/검색)
sentence-transformers
chromadb
torch # (sentence-transformers가 의존하므로 유지)

# DB 및 유틸리티
mysql-connector-python # cache.py에서 사용
requests
numpy
Pillow
python-dotenv